+++ cucumber-pdf-report/src/main/java/tech/grasshopper/pdf/annotation/AnnotationProcessor.java
@@ -1,14 +1,9 @@
 package tech.grasshopper.pdf.annotation;
 
-import org.apache.pdfbox.pdmodel.PDPage;
-import org.apache.pdfbox.pdmodel.common.PDRectangle;
 import org.apache.pdfbox.pdmodel.interactive.action.PDActionGoTo;
 import org.apache.pdfbox.pdmodel.interactive.annotation.PDAnnotationLink;
-import org.apache.pdfbox.pdmodel.interactive.annotation.PDBorderStyleDictionary;
 
 import lombok.Builder;
-import lombok.Getter;
-import lombok.Setter;
 import lombok.SneakyThrows;
 import tech.grasshopper.pdf.data.ReportData;
 import tech.grasshopper.pdf.destination.Destination;
@@ -16,33 +11,13 @@ import tech.grasshopper.pdf.pojo.cucumber.Executable;
 import tech.grasshopper.pdf.pojo.cucumber.Feature;
 import tech.grasshopper.pdf.pojo.cucumber.Scenario;
 
-@Setter
 @Builder
-public class Annotation {
+public class AnnotationProcessor {
 
-	@SuppressWarnings("unused")
-	private String title;
-	private float xBottom;
-	private float yBottom;
-	private float width;
-	private float height;
-	@Getter
-	private PDPage page;
-
-	public PDAnnotationLink createPDAnnotationLink() {
-
-		PDRectangle position = new PDRectangle(xBottom, yBottom, width, height);
-		PDAnnotationLink link = new PDAnnotationLink();
-
-		PDBorderStyleDictionary borderULine = new PDBorderStyleDictionary();
-		link.setBorderStyle(borderULine);
-
-		link.setRectangle(position);
-		return link;
-	}
+	private ReportData reportData;
 
 	@SneakyThrows
-	public static void updateDestination(ReportData reportData) {
+	public void updateDestination() {
 
 		for (Feature feature : reportData.getFeatures()) {
 			feature.getAnnotations().forEach(a -> {
@@ -55,9 +30,8 @@ public class Annotation {
 				});
 
 				for (Executable executable : scenario.getStepsAndHooks()) {
-					for (int i = 0; i < executable.getMediaAnnotations().size(); i++) {
-						updateDestination(executable.getMediaAnnotations().get(i),
-								executable.getMediaDestinations().get(i));
+					for (int i = 0; i < executable.getAnnotations().size(); i++) {
+						updateDestination(executable.getAnnotations().get(i), executable.getDestinations().get(i));
 					}
 				}
 			}
@@ -65,7 +39,7 @@ public class Annotation {
 	}
 
 	@SneakyThrows
-	private static void updateDestination(Annotation annotation, Destination destination) {
+	private void updateDestination(Annotation annotation, Destination destination) {
 
 		if (annotation == null || destination == null)
 			return;