+++ cucumber-pdf-report/src/main/java/tech/grasshopper/pdf/config/ReportConfig.java
@@ -13,14 +13,17 @@ public class ReportConfig {
 	private String passColor;
 	private String failColor;
 	private String skipColor;
+
 	private boolean displayFeature = true;
 	private boolean displayScenario = true;
 	private boolean displayDetailed = true;
-	private boolean displayExpanded = true;
+	private boolean displayExpanded = false;
+	private boolean displayAttached = true;
 
 	private SummaryConfig summaryConfig = new SummaryConfig();
 	private FeatureConfig featureConfig = new FeatureConfig();
 	private ScenarioConfig scenarioConfig = new ScenarioConfig();
+
 	private DetailedFeatureConfig detailedFeatureConfig = new DetailedFeatureConfig();
 	private DetailedScenarioConfig detailedScenarioConfig = new DetailedScenarioConfig();
 	private DetailedStepHookConfig detailedStepHookConfig = new DetailedStepHookConfig();